[201] Updated SD2. Updated scripts_texts and waypoints sqls. Never show standing with factions of the opposing team. Minor cleanup in affected code. Log contents of unprocessed tail data if logging is set to debug.
Remove watch flag (1) from characters that are watching a faction from the opposing team.
These factions will no longer be flagged for watching if they are from the opposing team. All we need to do is clean up the data from when it was still bugged.
NOTE: running this may take a while depending on the size of your character database.

Fix bug with end of winterpgrasp, as zone is cleaned of corpses we must revive players if dead.
Bitwise logic does not work on values that are not bitmasks. Thanks Spp for pointing this block of code out :)

Select only expired auctions when updating, by Paradox
Remove some useless code from AHBot, by Paradox